requirements.txt
streamlit>=1.28.0
whisper>=1.0.0
openai>=1.3.0
weave>=0.51.0
sounddevice>=0.4.6
numpy>=1.24.0
pandas>=2.0.0
jinja2>=3.1.2
wave>=0.0.2
requests>=2.31.0 
python-dotenv>=1.0.0